Separated REPL and file running logic into Outlet.CLI project
