--- Outlet/Properties/launchSettings.json
@@ -1,8 +1,0 @@
-{
-  "profiles": {
-    "Outlet": {
-      "commandName": "Project",
-      "workingDirectory": "C:\\Users\\coryl\\Programming Workspace\\C# Projects\\Outlet-Lang"
-    }
-  }
-}
